+++ mygeotab/dates.py
@@ -3,7 +3,6 @@
 from datetime import datetime
 
 import pytz
-from tzlocal import get_localzone
 
 
 def format_iso_datetime(dt):
@@ -17,16 +16,15 @@ def format_iso_datetime(dt):
     return dt.replace(tzinfo=None).strftime('%Y-%m-%dT%H:%M:%S.%f')[:-3] + 'Z'
 
 
-def localize_datetime(dt, tz=None):
+def localize_datetime(dt, tz=pytz.utc):
     """
-    Converts a naive or UTC-localized date into the provided timezone or current machine's timezone
+    Converts a naive or UTC-localized date into the provided timezone
 
     :rtype: datetime
     :param dt: The datetime object
-    :param tz: The timezone. If blank or None, the user's native timezone is used
+    :param tz: The timezone. If blank or None, UTC is used
     :return: The localized datetime object
     """
-    tz = tz or get_localzone()
     if not dt.tzinfo:
         return tz.localize(dt)
     else: