+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changes
 **Enhancements**
 
 - Deprecated the 'search()' function. Replaced by folding the previous functionality into 'get()'.
+- Removed 'tzlocal' dependency. Always deal with dates in UTC by default.
 
 
 0.4.4 (2016-07-10)